requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
psycopg2-binary==2.9.10
scikit-learn==1.5.2
mlflow==2.19.0
Flask==3.1.0
xgboost==2.1.3
metaflow==2.13